Dijkstra's Algorithm in Java
Dijkstra's algorithm is an algorithm for finding the shortest paths between a source node to all other nodes in a graph. We indirectly generate a SPT (shortest path tree) with a given source as a root. We maintain two sets, one set contains vertices included in the shortest-path tree, other set includes vertices not yet included in the shortest-path tree. At every step of the algorithm, we find a vertex that is in the other set (set of not yet included) and has a minimum distance from the source.